move parse logic in separate module
